Lab 2/requirements.txt
Django
psycopg2-binary
websockets
channels
djangorestframework